generate metadata
