+++ src/app/(home)/[categoryslug]/page.js
@@ -1,45 +1,18 @@
-"use client";
-import React, { useEffect, useState } from "react";
-import BlogPage from "@/components/blog/BlogPage";
-import Loading from "../loading";
+import CategorySlug from "./category-components/CategorySlug";
 
-const CategorySlug = ({ params }) => {
+export async function generateMetadata({ params, searchParams }, parent) {
   const { categoryslug } = params;
-  const [blogs, setBlogs] = useState([]);
-  const [loading, setLoading] = useState(true);
+  return {
+    title: categoryslug,
+  };
+}
 
-  useEffect(() => {
-    const fetchBlogsByCategory = async () => {
-      try {
-        const response = await fetch(
-          `/api/blogs/get-blogs?category=${categoryslug}`
-        );
-        const data = await response.json();
-
-        if (response.ok && Array.isArray(data.result)) {
-          setBlogs(data.result);
-        } else {
-          console.error("Error fetching blogs:", data.error);
-        }
-      } catch (error) {
-        console.error("Fetch error:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchBlogsByCategory();
-  }, [categoryslug]);
-
-  if (loading) {
-    return <Loading />;
-  }
-
+const CategorySlugPage = ({ params }) => {
   return (
     <main>
-      <BlogPage data={blogs} type="category" />
+      <CategorySlug params={params} />
     </main>
   );
 };
 
-export default CategorySlug;
+export default CategorySlugPage;
